requirements/base.txt
django-ninja==0.19.1
django-extensions==3.2.0
pyyaml==6.0